requirements.txt
torch==2.9.1
torchvision==0.24.1
numpy==2.4.0
matplotlib==3.10.8
tqdm>=4.65
PyYAML>=6.0
